requirements.txt
streamlit
pymongo
openai
python-dotenv
psycopg2-binary
rich
pydantic
pathlib
textwrap
typing-extensions
